Add XcvrBase as the common DeviceBase subclass for SfpBase and CpoBase, with shared transceiver API methods.
