+++ sonic_platform_base/__init__.py
@@ -7,6 +7,7 @@ from . import module_base
 from . import platform_base
 from . import psu_base
 from . import sfp_base
+from . import xcvr_base
 from . import thermal_base
 from . import sensor_base
 from . import watchdog_base
